backend/requirements.txt
spacy==3.2.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
matplotlib==3.5.1
Pillow==9.0.1
